Bugfix: fragmentation did not work correctly as the ID is the same for every detector so we have to add the source IP

+++ src/socket/FragmentStore.cpp
@@ -9,7 +9,7 @@
 
 namespace na62 {
 
-std::map<uint16_t, std::vector<DataContainer>> FragmentStore::fragmentsById_;
+std::map<uint64_t, std::vector<DataContainer>> FragmentStore::fragmentsById_;
 tbb::spin_mutex FragmentStore::newFragmentMutex_;
 
 uint FragmentStore::numberOfFragmentsReceived_ = 0;
